requirements.txt
gitdb==4.0.9
GitPython==3.1.27
PyYAML==6.0
semver==2.13.0
smmap==5.0.0
